Change labels dtype to int64 for cross entropy loss calculation

+++ dataloader.py
@@ -106,7 +106,7 @@ class TrafficVehicle(Dataset):
 
             img = torch.tensor(img, dtype=torch.float32).permute(2,0,1) #CHW format for PyTorch
             target['boxes'] = torch.tensor(target['boxes'], dtype=torch.float32)
-            target['labels'] = torch.tensor(target['labels'], dtype=torch.int32)
+            target['labels'] = torch.tensor(target['labels'], dtype=torch.int64)
         return img, target
